[#1745] Fix python linting errors
